Just messing around with HTML and CSS
Managed to float another <div> box in line with the other one, and put a
random button in it. The sizes are shit, and it looks bad, but it was
some good practice. So might as well commit it.

I'll probably try to get panes, and/or animations and/or new panes to
appear only after a certain point next.